==> 01_zveza_from_historic_court_cpp.ipynb <==
from glob import glob
import json
import re
import pandas as pd

zveza_all = []
for x in glob("./data/scraped/cpp_in__co_ali_go__in_vs/*.json"):
    with open(x, "r") as f:
        data = json.load(f)

    zveza_all.append((data["zveza"], data["meta"]["Evidenčna številka"]))


set([y for x in zveza_all for y in x[0].keys()])

zveza_unformatted = set([(*y, x[1]) for x in zveza_all for y in x[0].items()]) - set(
    [""]
)
zveza_unformatted = pd.DataFrame(
    zveza_unformatted, columns=["category", "zveza", "evidencna_stevilka"]
)
zveza_unformatted

# ZAKONI RS

def split_RS(zveza):
    zveza = re.sub(r"[\n\t\r]", "", zveza)

    split_phrase = "#$%^"
    pattern = r"[A-Z][\w\s]+\(\d+\)"
    # zveza_splitters = [".", "Konvencija ", "Direktiva ", "Uredba ", "Pravilnik "]

    matches = re.findall(pattern, zveza)
    matches.append("Zakon o")

    for match in matches:
        zveza = zveza.replace(match, f"{split_phrase}{match}")

    return [x.strip() for x in zveza.split(split_phrase) if x]


def _format_clen_from_docs(clen):
    if clen.isdigit():
        return f"{clen}. člen"
    digits = "".join([x for x in clen if x.isdigit()])
    alphas = "".join([x for x in clen if x.isalpha()])
    return f"{digits}.{alphas} člen"


formatted_data = []
for _val in zveza_unformatted.loc[
    zveza_unformatted["category"] == "RS - Ustava, Zakoni, Sporazumi, Pogodbe",
    ["zveza", "evidencna_stevilka"],
].values:
    zveza, evidencna_stevilka = _val
    for z in split_RS(zveza):
        try:
            _tmp = z.split(" - ")
            if len(_tmp) == 3:
                title, abbr, clens_unformatted = _tmp
            else:
                title, clens_unformatted = _tmp
                abbr = ""

            title, abbr, clens_unformatted = (
                title.strip(),
                abbr.strip(),
                clens_unformatted.strip(),
            )
            title = re.sub(r"\(\d+\)", "", title)
        except:
            print("ERROR split: ", z)
            continue

        try:
            zveza_true = []
            for clen in clens_unformatted.split(","):
                clen = clen.strip()
                clen = clen.split("/")[0]
                clen = clen.split("-")[0]
                if _split := clen.split(" "):
                    clen = [x for x in _split if not all(y.isalpha() for y in x)][0]
                zveza_true.append(clen)

            zveza_true = sorted(
                [f"{_format_clen_from_docs(x)} {abbr}" for x in list(set(zveza_true))],
                key=lambda x: int(x.split(".")[0]),
            )
            formatted_data.append(
                {
                    "title": title,
                    "abbr": abbr,
                    "zveza": zveza_true,
                    "evidencna_stevilka": evidencna_stevilka,
                }
            )
        except:
            print("ERROR formatting: ", zveza)
            continue


RS_cleaned_data = pd.DataFrame(formatted_data)
RS_cleaned_data

# WITHOUT CATEGORY

def split_uncatagorized(zveza):
    zveza = re.sub(r"[\n\t\r]", "", zveza)
    zveza = zveza.replace("ZPP (1977)", "ZPP")

    split_phrase = "#$%^"
    pattern = r"[\w\d-]+ člen"

    matches = re.findall(pattern, zveza)

    for match in matches:
        zveza = zveza.replace(match, f"{split_phrase}{match}")

    return [x.strip() for x in zveza.split(split_phrase) if x]


formatted_data = []
for x in zveza_unformatted.loc[
    zveza_unformatted["category"] == "Without category",
    ["zveza", "evidencna_stevilka"],
].values:
    zveza, evidencna_stevilka = x
    zveza = split_uncatagorized(zveza)

    for z in zveza:
        pattern = r"[\w\d-]+ člen"
        matches = re.findall(pattern, z)

        if len(matches) != 1:
            print("ERROR MATCH: ", z, matches)
            continue

        abbr = matches[0].replace("člen", "").strip()
        z = z.replace(matches[0], "")
        zveza_split = z.split(",")
        zveza_split = [
            x.strip().strip(".").split("/")[0].split("-")[0] for x in zveza_split if x
        ]

        try:
            zveza_true = sorted(
                [f"{_format_clen_from_docs(x)} {abbr}" for x in list(set(zveza_split))],
                key=lambda x: int(x.split(".")[0]),
            )
        except:
            print("ERROR SPLIT: ", zveza_split)
            continue

        formatted_data.append(
            {
                "title": "Without category",
                "abbr": abbr,
                "zveza": zveza_true,
                "evidencna_stevilka": evidencna_stevilka,
            }
        )
WC_cleaned_data = pd.DataFrame(formatted_data)
WC_cleaned_data


EU

eu_types = [
    "EU - Direktive, Uredbe, Sklepi / Odločbe, Sporazumi, Pravila",
]

formatted_data = []
for x in zveza_unformatted.loc[
    zveza_unformatted["category"].isin(eu_types),
    ["zveza", "evidencna_stevilka"],
].values:
    zveza, evidencna_stevilka = x
    if "Direktiva" in zveza and not "Uredba" in zveza:
        zveza, evidencna_stevilka = x

        for _direktiva in [
            f"Direktiva{x}" for x in zveza.split("Direktiva") if x.strip()
        ]:
            title, clens_unformatted = _direktiva.split(" - ")
            pattern = r"\d{1,2}/\d{1,2}/EGS"
            matches = re.findall(pattern, title)

            if len(matches) != 1:
                print("ERROR MATCH: ", _direktiva)
                continue
            abbr = matches[0]

            clens_unformatted = clens_unformatted.replace("člen", "")
            clens_split = [
                x.strip().strip(".").split("/")[0].split("-")[0]
                for x in clens_unformatted.split(",")
            ]

            try:
                zveza_true = sorted(
                    [
                        f"{_format_clen_from_docs(x)} {abbr}"
                        for x in list(set(clens_split))
                    ],
                    key=lambda x: int(x.split(".")[0]),
                )
            except:
                print("ERROR SPLIT: ", clens_split)
                continue

            formatted_data.append(
                {
                    "title": title,
                    "abbr": abbr,
                    "zveza": zveza_true,
                    "evidencna_stevilka": evidencna_stevilka,
                }
            )
EU_cleaned_data = pd.DataFrame(formatted_data)
EU_cleaned_data


eu_types = [
    "EU - Direktive, Uredbe, Sklepi / Odločbe, Sporazumi, Pravila",
]

formatted_data = []
for x in zveza_unformatted.loc[
    zveza_unformatted["category"].isin(eu_types),
    ["zveza", "evidencna_stevilka"],
].values:
    zveza, evidencna_stevilka = x
    if "Uredba" in zveza and not "Direktiva" in zveza:
        zveza, evidencna_stevilka = x

        for _uredba in [f"Uredba{x}" for x in zveza.split("Uredba") if x.strip()]:
            print(_uredba)
            title, clens_unformatted = _uredba.split(" - ")
            pattern = r"\d{3,4}/\d{4}"
            matches = re.findall(pattern, title)

            if not matches:
                pattern = r"\d{1,2}/\d{4}"
                matches = re.findall(pattern, title)

            if len(matches) != 1:
                print("ERROR MATCH: ", _uredba)
                continue
            abbr = matches[0]

            clens_unformatted = clens_unformatted.replace("člen", "")
            clens_split = [
                x.strip().strip(".").split("/")[0].split("-")[0]
                for x in clens_unformatted.split(",")
            ]

            try:
                zveza_true = sorted(
                    [
                        f"{_format_clen_from_docs(x)} {abbr}"
                        for x in list(set(clens_split))
                    ],
                    key=lambda x: int(x.split(".")[0]),
                )
            except:
                print("ERROR SPLIT: ", clens_split)
                continue

            formatted_data.append(
                {
                    "title": title,
                    "abbr": abbr,
                    "zveza": zveza_true,
                    "evidencna_stevilka": evidencna_stevilka,
                }
            )
formatted_data = pd.DataFrame(formatted_data)
formatted_data


OTHER

other_types = [
    "Konvencije, Deklaracije Resolucije",
    "Mednarodne Pogodbe",
    "Podzakonski akti / Vsi drugi akti",
    "Predpisi drugih držav",
]

formatted_data = []
for x in zveza_unformatted.loc[
    zveza_unformatted["category"].isin(other_types),
    ["zveza", "evidencna_stevilka"],
].values:
    zveza, evidencna_stevilka = x
    print(zveza)


merged_data = pd.concat([RS_cleaned_data, WC_cleaned_data, EU_cleaned_data])
merged_data["title"] = merged_data["title"].str.replace("Without category", "")
merged_data["title"] = merged_data["title"].str.strip()

# clean wrongly processed data
# TODO: no time to debug this
merged_data = merged_data[
    ~(
        merged_data["abbr"].isin(
            [
                "zadevah",
                "red",
                "pravicah",
                "izletih",
                "cest",
                "občine",
                "tarifa",
                "Slovenije",
                "premoženja",
                "vprašanjih",
                "energije",
                "zemljišča",
                "gumami",
                "nesreče",
                "cestah",
                "Hrvaške",
                "cenilcih",
                "razlastitvi",
                "1974",
                "vodah",
                "svoboščin",
                "zavarovanja",
                "99",
                "otroki",
                "1977",
                "pokojnin",
                "pravic",
                "mestih",
                "stanovanj",
                "1",
                "postopku",
                "poročila",
                "opreme",
                "NE-91",
                "NE-zahtevkov88",
                "A",
                "najemnin",
                "storitev",
                "fonogramov",
                "stavb",
                "delom",
                "dela",
                "SODIŠČA",
                "RED",
                "delovanje",
                "okupacijo",
                "zemljiščem",
                "park",
                "sodiščih",
                "toka",
                "etike",
                "skupnostih",
                "POSTOPKU",
                "95",
                "1996",
                "Ustava",
                "PRAVICAH",
                "konvencija",
                "odhodku",
                "članicah",
                "0Z",
                "znamkami",
                "ZDoh-",
                "okolju",
                "terjatev",
                "sistemu",
                "preoblikovanja",
                "denacionalizacije",
                "gozdarstvu",
                "naročil",
                "civile",
                "otrok",
                "države",
                "zneskov",
                "register",
                "naročanju",
                "P-93",
                "pravila",
                "pogodba",
                "Skupnosti",
                "sodišč",
                "postopka",
                "1989",
                "delu",
                "občin",
                "zahtevkov",
                "blaga",
            ]
        )
    )
]


merged_data.reset_index(drop=True).to_parquet(
    "./data/mappings/legislation_mappings_from_cpp_scraped.parquet"
)
