==> 02_kzp_cpp_mapping_analysis.ipynb <==
import pandas as pd
from src.nlp.text_processor import TextProcessor

tp = TextProcessor()

kzp_mappings = pd.read_parquet(
    "./data/mappings/legislation_mappings_from_kzp_scraped.parquet"
)
cpp_mappings = pd.read_parquet(
    "./data/mappings/legislation_mappings_from_cpp_scraped.parquet"
)
existing_mappings = pd.read_parquet(
    "./data/mappings/legislation_mappings_from_db.parquet"
)

kzp_mappings["abbr"].value_counts()[:20]

cpp_mappings["abbr"].value_counts()[:20]

scraped_mappings = (
    pd.concat([kzp_mappings, cpp_mappings])
    .drop_duplicates(subset=["title", "abbr"])[["title", "abbr"]]
    .reset_index(drop=True)
)

court_mappings_processed = []
for _, row in scraped_mappings.iterrows():
    _abbr = row["abbr"]
    title = row["title"]
    if _abbr and title:
        title += f" ({_abbr})"

    court_mappings_processed.append(
        {
            "document_title": title,
            "document_title_striped": row["title"],
            "document_title_lemmatized": tp.lemmatize_text(row["title"]),
            "abbreviation": _abbr,
        }
    )

court_mappings_processed = pd.DataFrame(court_mappings_processed)


court_mappings_processed

mappings_all = pd.concat([existing_mappings, court_mappings_processed]).drop_duplicates(
    subset=["document_title", "abbreviation"]
)
mappings_all = mappings_all[
    (mappings_all["document_title_striped"].str.len() > 0)
    & (mappings_all["document_title_striped"].str.len() > 0)
]

mappings_all.to_parquet("./data/mappings/mappings_all.parquet")
