==> 02_mappings_processor_db.ipynb <==
import pandas as pd
from src.parse import _get_law_abbreviation, _strip_law_abbreviation
from src.nlp.text_processor import TextProcessor

tp = TextProcessor()

laws_staging = pd.read_parquet("./data/mappings/legislation_laws_staging.parquet")
laws_staging.head()

_tmp = laws_staging[:1]
_tmp["document_title"] = "Ustava"
_tmp["document_title_lemmatized"] = "ustava"
_tmp

laws_staging = pd.concat([_tmp, laws_staging]).reset_index(drop=True)
laws_staging

laws_staging.shape

docs = pd.read_parquet("./data/mappings/document_overview.parquet")
docs.head()

new_db_abbrs = (
    docs[docs["document_type_class"] == "zakon_tfl"]
    .apply(lambda x: _get_law_abbreviation(x["document_title"]), axis=1)
    .to_list()
)

new_db_titles = docs.loc[
    docs["document_type_class"] == "zakon_tfl", "document_title"
].to_list()

new_db_titles_stripped = (
    docs[docs["document_type_class"] == "zakon_tfl"]
    .apply(lambda x: _strip_law_abbreviation(x["document_title"]), axis=1)
    .to_list()
)

new_db_titles_lemmatized = [tp.lemmatize_text(x) for x in new_db_titles_stripped]


new_entries = pd.DataFrame.from_records(
    [
        {
            "document_title": title,
            "abbreviation": abbr,
            "document_title_striped": stripped,
            "document_title_lemmatized": lemma,
        }
        for title, abbr, stripped, lemma in zip(
            new_db_titles,
            new_db_abbrs,
            new_db_titles_stripped,
            new_db_titles_lemmatized,
        )
    ]
)

pd.concat([laws_staging, new_entries])[
    [
        "document_title",
        "document_title_striped",
        "document_title_lemmatized",
        "abbreviation",
    ]
].drop_duplicates().sort_values(by="document_title").to_parquet(
    "./data/mappings/legislation_mappings_from_db.parquet"
)
