==> 03_create_training_dataset.ipynb <==
import pandas as pd
from glob import glob
import json
import os

kzp_mappings = pd.read_parquet(
    "./data/mappings/legislation_mappings_from_kzp_scraped.parquet"
)
cpp_mappings = pd.read_parquet(
    "./data/mappings/legislation_mappings_from_cpp_scraped.parquet"
)

scraped_mappings = pd.concat([kzp_mappings, cpp_mappings]).reset_index(drop=True)
scraped_mappings


for evidencna_stevilka, _group in scraped_mappings.groupby("evidencna_stevilka"):
    zakonodaja_naslovi = sorted(_group["title"].values.tolist())
    zakonadaja_kratice = sorted(_group["abbr"].values.tolist())

    zveza = []
    for x in _group["zveza"].values.tolist():
        zveza.extend(x)
    zveza = sorted(zveza)

    files = glob(f"./data/scraped/*/{evidencna_stevilka}.json")
    assert len(files) == 1

    with open(files[0], "r") as f:
        data = json.load(f)

    mapping = {
        "meta": data["meta"],
        "content": data["content"],
        "processed": {
            "zakonodaja_naslovi": zakonodaja_naslovi,
            "zakonadaja_kratice": zakonadaja_kratice,
            "zveza": zveza,
            "evidencna_stevilka": evidencna_stevilka,
        },
    }

    os.makedirs("./data/training", exist_ok=True)
    with open(f"./data/training/{evidencna_stevilka}.json", "w") as f:
        json.dump(mapping, f, indent=4)
